Another Commit

+++ base/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
-from .models import Room, Topic
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 
@@ -17,16 +20,48 @@ from .forms import RoomForm
 
 
 def loginPage(request):
-    context = {}
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
         except:
             messages.error(request, "User Does Not Exist")
 
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password Does not Exist')
+    context = {'page':page}
+
     return render(request, 'base/login_register.html', context)
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('login')
+        else:
+            messages.error(request, 'An Error Occured During Registration')
+    return render(request, 'base/login_register.html', {'form':form})
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -38,14 +73,34 @@ def home(request):
 
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    print(room_messages)
+    context = {'rooms':rooms, 'topics':topics, 'room_count':room_count, ' room_messages':room_messages}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages = room.message_set.all()
+    participants = room.participant.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participant.add(request.user)
+    
+        return redirect('room', pk=room.id)
+    context = {'room':room, 'room_messages':room_messages, 'participants':participants}
     return render(request, 'base/room.html', context)
 
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    context = {'user':user}
+    return render(request, 'base/profile.html', context)
+
+@login_required(login_url='login') 
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -56,9 +111,16 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You Are Not Allowed Here!!') 
+
+    if request.user != room.host:
+        return HttpResponse('You Are Not Allowed Here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -69,12 +131,26 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You Are Not Allowed Here!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You Are Not Allowed Here!!')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
